pyproject.toml
[project]
name = "nonebot-plugin-nobot-here"
version = "0.1.0"
description = "哪里有机器人?咱不是机器人!"
authors = [
    {name = "Cvandia", email = "106718176+Cvandia@users.noreply.github.com"},
]
dependencies = ["nonebot2>=2.2.0", "nonebot-plugin-alconna>=0.54.0"]
requires-python = ">=3.10,<4.0"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = true

[dependency-groups]
dev = ["pre-commit>=4.0.1", "ruff>=0.8.0"]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
